@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "cognite-sdk"
 
-version = "7.41.0"
+version = "7.53.2"
 description = "Cognite Python SDK"
 readme = "README.md"
 documentation = "https://cognite-sdk-python.readthedocs-hosted.com"
@@ -30,27 +30,37 @@ msal = "^1"
 protobuf = ">=4"
 pip = ">=20.0.0"  # make optional once poetry doesn't auto-remove it on "simple install"
 typing_extensions = ">= 4"
-numpy = { version = "^1.20", optional = true }
+backports-zoneinfo = { version = ">=0.2.1",  python = "<3.9" }
+# Windows does not have a ANSI database and need tzdata... pyodide also needs it:
+tzdata = { version = ">=2024.1", markers = "platform_system == 'Windows' or platform_system == 'Emscripten'" }
+numpy = [
+    { version = ">=1.20, <1.25", python = "~3.8", optional = true },
+    { version = "^1.25", python = ">=3.9, <3.12", optional = true },
+    { version = "^1.26", python = "^3.12", optional = true },
+]
 sympy = { version = "*", optional = true }
-pandas = { version = ">=1.4", optional = true }
-backports-zoneinfo = { version = ">=0.2.1",  python = "<3.9", extras = ["tzdata"], optional = true}
-# Windows does not have a ANSI database and need tzdata
-tzdata = {version = ">=2023.3", markers = "sys_platform == 'win32'", optional = true }
-geopandas = { version = ">=0.10.0", optional = true }
+pandas = [
+    { version = ">=1.4, <2.1", python = "~3.8", optional = true },
+    { version = ">=2.1", python = ">=3.9", optional = true },
+]
+geopandas = [
+    { version = ">=0.10, <0.14", python = "~3.8", optional = true },
+    { version = ">=0.14", python = ">=3.9", optional = true },
+]
 shapely = { version = ">=1.7.0", optional = true }
-pyodide-http = {version = "^0.2.1", optional = true }
-graphlib-backport = {version = "^1.0.0", python = "<3.9"}
-PyYAML = {version = "^6.0", optional = true}
+pyodide-http = { version = "^0.2.1", optional = true }
+graphlib-backport = { version = "^1.0.0", python = "<3.9" }
+PyYAML = { version = "^6.0", optional = true }
 
 [tool.poetry.extras]
-pandas = ["pandas", "backports-zoneinfo"]
+pandas = ["pandas"]
 numpy = ["numpy"]
 geo = ["geopandas", "shapely"]
 sympy = ["sympy"]
 functions = ["pip"]
 yaml = ["PyYAML"]
-pyodide = ["pyodide-http"]  # keep pyodide related dependencies outside of 'all'
-all = ["numpy", "pandas", "geopandas", "shapely", "sympy", "pip", "backports-zoneinfo", "PyYAML"]
+pyodide = ["pyodide-http", "tzdata"]  # keep pyodide related dependencies outside of 'all'
+all = ["numpy", "pandas", "geopandas", "shapely", "sympy", "pip", "PyYAML"]
 
 [tool.poetry.group.dev.dependencies]
 docutils = "==0.15.2"
@@ -62,7 +72,7 @@ pytest-rerunfailures = ">=10"
 pytest-asyncio = ">=0.23.5a0"
 pytest-xdist = ">=2"
 twine = "^4.0.1"
-pre-commit = "^2.20.0"
+pre-commit = "^3"
 responses = "^0"
 toml = "^0"
 python-dotenv = "^0"
