@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from abc import ABC
-from typing import TYPE_CHECKING, Any, Literal, MutableSequence, Sequence, Tuple, Union
+from collections.abc import Iterator
+from typing import TYPE_CHECKING, Any, Literal, MutableSequence, Tuple, Union, overload
 from urllib.parse import quote
 
 from typing_extensions import TypeAlias
@@ -9,13 +9,13 @@ from typing_extensions import TypeAlias
 from cognite.client._api_client import APIClient
 from cognite.client._constants import DEFAULT_LIMIT_READ
 from cognite.client.data_classes.workflows import (
-    CancelExecution,
     Workflow,
     WorkflowExecution,
     WorkflowExecutionDetailed,
     WorkflowExecutionList,
     WorkflowIds,
     WorkflowList,
+    WorkflowStatus,
     WorkflowTaskExecution,
     WorkflowUpsert,
     WorkflowVersion,
@@ -24,7 +24,6 @@ from cognite.client.data_classes.workflows import (
     WorkflowVersionUpsert,
 )
 from cognite.client.exceptions import CogniteAPIError
-from cognite.client.utils._experimental import FeaturePreviewWarning
 from cognite.client.utils._identifier import (
     IdentifierSequence,
     WorkflowVersionIdentifierSequence,
@@ -36,32 +35,17 @@ if TYPE_CHECKING:
     from cognite.client import ClientConfig, CogniteClient
     from cognite.client.data_classes import ClientCredentials
 
-
-class BetaWorkflowAPIClient(APIClient, ABC):
-    def __init__(
-        self,
-        config: ClientConfig,
-        api_version: str | None,
-        cognite_client: CogniteClient,
-    ) -> None:
-        super().__init__(config, api_version, cognite_client)
-        self._api_subversion = "beta"
-        self._warning = FeaturePreviewWarning(
-            api_maturity="beta", sdk_maturity="alpha", feature_name="Workflow Orchestration"
-        )
-
-
 WorkflowIdentifier: TypeAlias = Union[WorkflowVersionId, Tuple[str, str], str]
 WorkflowVersionIdentifier: TypeAlias = Union[WorkflowVersionId, Tuple[str, str]]
 
 
-class WorkflowTaskAPI(BetaWorkflowAPIClient):
+class WorkflowTaskAPI(APIClient):
     _RESOURCE_PATH = "/workflows/tasks"
 
     def update(
         self, task_id: str, status: Literal["completed", "failed"], output: dict | None = None
     ) -> WorkflowTaskExecution:
-        """`Update status of async task. <https://api-docs.cognite.com/20230101-beta/tag/Tasks/operation/UpdateTaskStatus>`_
+        """`Update status of async task. <https://api-docs.cognite.com/20230101/tag/Tasks/operation/UpdateTaskStatus>`_
 
         For tasks that has been marked with 'is_async = True', the status must be updated by calling this endpoint with either 'completed' or 'failed'.
 
@@ -96,8 +80,6 @@ class WorkflowTaskAPI(BetaWorkflowAPIClient):
                 >>> res = client.workflows.tasks.update(res.tasks[1].id, "completed")
 
         """
-        self._warning.warn()
-
         body: dict[str, Any] = {"status": status.upper()}
         if output is not None:
             body["output"] = output
@@ -108,11 +90,11 @@ class WorkflowTaskAPI(BetaWorkflowAPIClient):
         return WorkflowTaskExecution.load(response.json())
 
 
-class WorkflowExecutionAPI(BetaWorkflowAPIClient):
+class WorkflowExecutionAPI(APIClient):
     _RESOURCE_PATH = "/workflows/executions"
 
     def retrieve_detailed(self, id: str) -> WorkflowExecutionDetailed | None:
-        """`Retrieve a workflow execution with detailed information. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-executions/operation/ExecutionOfSpecificRunOfWorkflow>`_
+        """`Retrieve a workflow execution with detailed information. <https://api-docs.cognite.com/20230101/tag/Workflow-executions/operation/ExecutionOfSpecificRunOfWorkflow>`_
 
         Args:
             id (str): The server-generated id of the workflow execution.
@@ -136,7 +118,6 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
                 >>> res = client.workflows.executions.retrieve_detailed(res[0].id)
 
         """
-        self._warning.warn()
         try:
             response = self._get(url_path=f"{self._RESOURCE_PATH}/{id}")
         except CogniteAPIError as e:
@@ -153,7 +134,7 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
         metadata: dict | None = None,
         client_credentials: ClientCredentials | None = None,
     ) -> WorkflowExecution:
-        """`Trigger a workflow execution. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-executions/operation/TriggerRunOfSpecificVersionOfWorkflow>`_
+        """`Trigger a workflow execution. <https://api-docs.cognite.com/20230101/tag/Workflow-executions/operation/TriggerRunOfSpecificVersionOfWorkflow>`_
 
         Args:
             workflow_external_id (str): External id of the workflow.
@@ -195,7 +176,6 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
                 >>> credentials = ClientCredentials("my-client-id", os.environ["MY_CLIENT_SECRET"])
                 >>> res = client.workflows.executions.trigger("foo", "1", client_credentials=credentials)
         """
-        self._warning.warn()
         nonce = create_session_and_return_nonce(
             self._cognite_client, api_name="Workflow API", client_credentials=client_credentials
         )
@@ -213,17 +193,18 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
         workflow_version_ids: WorkflowVersionIdentifier | MutableSequence[WorkflowVersionIdentifier] | None = None,
         created_time_start: int | None = None,
         created_time_end: int | None = None,
+        statuses: WorkflowStatus | MutableSequence[WorkflowStatus] | None = None,
         limit: int = DEFAULT_LIMIT_READ,
     ) -> WorkflowExecutionList:
-        """`List workflow executions in the project. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-executions/operation/ListWorkflowExecutions>`_
+        """`List workflow executions in the project. <https://api-docs.cognite.com/20230101/tag/Workflow-executions/operation/ListWorkflowExecutions>`_
 
         Args:
             workflow_version_ids (WorkflowVersionIdentifier | MutableSequence[WorkflowVersionIdentifier] | None): Workflow version id or list of workflow version ids to filter on.
             created_time_start (int | None): Filter out executions that was created before this time. Time is in milliseconds since epoch.
             created_time_end (int | None): Filter out executions that was created after this time. Time is in milliseconds since epoch.
+            statuses (WorkflowStatus | MutableSequence[WorkflowStatus] | None): Workflow status or list of workflow statuses to filter on.
             limit (int): Maximum number of results to return. Defaults to 25. Set to -1, float("inf") or None
                         to return all items.
-
         Returns:
             WorkflowExecutionList: The requested workflow executions.
 
@@ -243,7 +224,6 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
                 >>> res = client.workflows.executions.list(created_time_start=int((datetime.now() - timedelta(days=1)).timestamp() * 1000))
 
         """
-        self._warning.warn()
         filter_: dict[str, Any] = {}
         if workflow_version_ids is not None:
             filter_["workflowFilters"] = WorkflowIds.load(workflow_version_ids).dump(
@@ -253,6 +233,11 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
             filter_["createdTimeStart"] = created_time_start
         if created_time_end is not None:
             filter_["createdTimeEnd"] = created_time_end
+        if statuses is not None:
+            if isinstance(statuses, MutableSequence):
+                filter_["status"] = [status.upper() for status in statuses]
+            else:  # Assume it is a stringy type
+                filter_["status"] = [statuses.upper()]
 
         return self._list(
             method="POST",
@@ -262,39 +247,42 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
             limit=limit,
         )
 
-    def cancel(self, executions: Sequence[CancelExecution]) -> Sequence[WorkflowExecution]:
-        """`cancel a workflow execution. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-executions/operation/WorkflowExecutionCancellation>`_
+    def cancel(self, id: str, reason: str | None) -> WorkflowExecution:
+        """`cancel a workflow execution. <https://api-docs.cognite.com/20230101/tag/Workflow-executions/operation/WorkflowExecutionCancellation>`_
 
         Args:
-            executions (Sequence[CancelExecution]): List of executions to cancel.
+            id (str): The server-generated id of the workflow execution.
+            reason (str | None): The reason for the cancellation, this will be put within the execution's `reasonForIncompletion` field. It is defaulted to 'cancelled' if not provided.
 
-        Tip:
-            Cancelling a workflow only prevents it from starting new tasks, tasks already running on
-            other services (transformations and functions) will keep running there.
+        Note:
+            Cancelling a workflow will immediately cancel the `in_progress` tasks, but not their spawned work in
+            other services (like transformations and functions).
 
         Returns:
-            Sequence[WorkflowExecution]: The canceled workflow executions.
+            WorkflowExecution: The canceled workflow execution.
 
         Examples:
 
             Trigger a workflow execution for the workflow "foo", version 1 and cancel it:
 
                 >>> from cognite.client import CogniteClient
-                >>> from cognite.client.data_classes import CancelExecution
                 >>> client = CogniteClient()
                 >>> res = client.workflows.executions.trigger("foo", "1")
-                >>> client.workflows.executions.cancel([CancelExecution(res.id, "test cancelation")])
+                >>> client.workflows.executions.cancel(id="foo", reason="test cancelation")
         """
-        self._warning.warn()
         response = self._post(
-            url_path=f"{self._RESOURCE_PATH}/cancel",
-            json={"items": [execution.dump(camel_case=True) for execution in executions]},
+            url_path=f"{self._RESOURCE_PATH}/{id}/cancel",
+            json={
+                "reason": reason,
+            }
+            if reason
+            else {},
         )
 
-        return [WorkflowExecution._load(execution) for execution in response.json()["items"]]
+        return WorkflowExecution._load(response.json())
 
     def retry(self, id: str, client_credentials: ClientCredentials | None = None) -> WorkflowExecution:
-        """`Retry a workflow execution. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-executions/operation/WorkflowExecutionRetryn>`_
+        """`Retry a workflow execution. <https://api-docs.cognite.com/20230101/tag/Workflow-executions/operation/WorkflowExecutionRetryn>`_
 
         Args:
             id (str): The server-generated id of the workflow execution.
@@ -307,13 +295,11 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
             Retry a workflow execution that has been cancelled or failed:
 
                 >>> from cognite.client import CogniteClient
-                >>> from cognite.client.data_classes import CancelExecution
                 >>> client = CogniteClient()
                 >>> res = client.workflows.executions.trigger("foo", "1")
-                >>> client.workflows.executions.cancel([CancelExecution(res.id, "test cancellation")])
+                >>> client.workflows.executions.cancel(id=res.id, reason="test cancellation")
                 >>> client.workflows.executions.retry(res.id)
         """
-        self._warning.warn()
         nonce = create_session_and_return_nonce(
             self._cognite_client, api_name="Workflow API", client_credentials=client_credentials
         )
@@ -324,15 +310,56 @@ class WorkflowExecutionAPI(BetaWorkflowAPIClient):
         return WorkflowExecution._load(response.json())
 
 
-class WorkflowVersionAPI(BetaWorkflowAPIClient):
+class WorkflowVersionAPI(APIClient):
     _RESOURCE_PATH = "/workflows/versions"
 
     def __init__(self, config: ClientConfig, api_version: str | None, cognite_client: CogniteClient) -> None:
         super().__init__(config, api_version, cognite_client)
         self._DELETE_LIMIT = 100
 
+    @overload
+    def __call__(
+        self,
+        chunk_size: None = None,
+        workflow_version_ids: WorkflowIdentifier | MutableSequence[WorkflowIdentifier] | None = None,
+        limit: int | None = None,
+    ) -> Iterator[WorkflowVersion]: ...
+
+    @overload
+    def __call__(
+        self,
+        chunk_size: int,
+        workflow_version_ids: WorkflowIdentifier | MutableSequence[WorkflowIdentifier] | None = None,
+        limit: int | None = None,
+    ) -> Iterator[WorkflowVersionList]: ...
+
+    def __call__(
+        self,
+        chunk_size: int | None = None,
+        workflow_version_ids: WorkflowIdentifier | MutableSequence[WorkflowIdentifier] | None = None,
+        limit: int | None = None,
+    ) -> Iterator[WorkflowVersion] | Iterator[WorkflowVersionList]:
+        """Iterate over workflow versions
+
+        Args:
+            chunk_size (int | None): The number of workflow versions to return in each chunk. Defaults to yielding one workflow version at a time.
+            workflow_version_ids (WorkflowIdentifier | MutableSequence[WorkflowIdentifier] | None): Workflow version id or list of workflow version ids to filter on.
+            limit (int | None): Maximum number of workflow versions to return. Defaults to returning all.
+
+        Returns:
+            Iterator[WorkflowVersion] | Iterator[WorkflowVersionList]: Returns an iterator over workflow versions.
+
+        """
+        return self._list_generator(
+            method="GET", resource_cls=WorkflowVersion, list_cls=WorkflowVersionList, limit=limit, chunk_size=chunk_size
+        )
+
+    def __iter__(self) -> Iterator[WorkflowVersion]:
+        """Iterate all over workflow versions"""
+        return self.__call__()
+
     def upsert(self, version: WorkflowVersionUpsert, mode: Literal["replace"] = "replace") -> WorkflowVersion:
-        """`Create a workflow version. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-versions/operation/CreateOrUpdateWorkflowVersion>`_
+        """`Create a workflow version. <https://api-docs.cognite.com/20230101/tag/Workflow-versions/operation/CreateOrUpdateWorkflowVersion>`_
 
         Note this is an upsert endpoint, so if a workflow with the same version external id already exists, it will be updated.
 
@@ -370,7 +397,6 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
                 ... )
                 >>> res = client.workflows.versions.upsert(new_version)
         """
-        self._warning.warn()
         if mode != "replace":
             raise ValueError("Only replace mode is supported for upserting workflow versions.")
 
@@ -386,7 +412,7 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
         workflow_version_id: WorkflowVersionIdentifier | MutableSequence[WorkflowVersionIdentifier],
         ignore_unknown_ids: bool = False,
     ) -> None:
-        """`Delete a workflow version(s). <https://api-docs.cognite.com/20230101-beta/tag/Workflow-versions/operation/DeleteSpecificVersionsOfWorkflow>`_
+        """`Delete a workflow version(s). <https://api-docs.cognite.com/20230101/tag/Workflow-versions/operation/DeleteSpecificVersionsOfWorkflow>`_
 
         Args:
             workflow_version_id (WorkflowVersionIdentifier | MutableSequence[WorkflowVersionIdentifier]): Workflow version id or list of workflow version ids to delete.
@@ -408,7 +434,6 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
                 >>> client.workflows.versions.delete([WorkflowVersionId("my workflow", "1"), WorkflowVersionId("my workflow 2", "2")])
 
         """
-        self._warning.warn()
         identifiers = WorkflowIds.load(workflow_version_id).dump(camel_case=True)
         self._delete_multiple(
             identifiers=WorkflowVersionIdentifierSequence.load(identifiers),
@@ -417,7 +442,7 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
         )
 
     def retrieve(self, workflow_external_id: str, version: str) -> WorkflowVersion | None:
-        """`Retrieve a workflow version. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-versions/operation/GetSpecificVersion>`_
+        """`Retrieve a workflow version. <https://api-docs.cognite.com/20230101/tag/Workflow-versions/operation/GetSpecificVersion>`_
 
         Args:
             workflow_external_id (str): External id of the workflow.
@@ -434,7 +459,6 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
                 >>> client = CogniteClient()
                 >>> res = client.workflows.versions.retrieve("my workflow", "1")
         """
-        self._warning.warn()
         try:
             response = self._get(
                 url_path=f"/workflows/{quote(workflow_external_id, '')}/versions/{quote(version, '')}",
@@ -451,7 +475,7 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
         workflow_version_ids: WorkflowIdentifier | MutableSequence[WorkflowIdentifier] | None = None,
         limit: int = DEFAULT_LIMIT_READ,
     ) -> WorkflowVersionList:
-        """`List workflow versions in the project <https://api-docs.cognite.com/20230101-beta/tag/Workflow-versions/operation/ListWorkflowVersions>`_
+        """`List workflow versions in the project <https://api-docs.cognite.com/20230101/tag/Workflow-versions/operation/ListWorkflowVersions>`_
 
         Args:
             workflow_version_ids (WorkflowIdentifier | MutableSequence[WorkflowIdentifier] | None): Workflow version id or list of workflow version ids to filter on.
@@ -482,7 +506,6 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
                 >>> res = client.workflows.versions.list([("my_workflow", "1"), ("my_workflow_2", "2")])
 
         """
-        self._warning.warn()
         if workflow_version_ids is None:
             workflow_ids_dumped = []
         else:
@@ -497,7 +520,7 @@ class WorkflowVersionAPI(BetaWorkflowAPIClient):
         )
 
 
-class WorkflowAPI(BetaWorkflowAPIClient):
+class WorkflowAPI(APIClient):
     _RESOURCE_PATH = "/workflows"
 
     def __init__(
@@ -512,8 +535,35 @@ class WorkflowAPI(BetaWorkflowAPIClient):
         self.tasks = WorkflowTaskAPI(config, api_version, cognite_client)
         self._DELETE_LIMIT = 100
 
+    @overload
+    def __call__(self, chunk_size: None = None, limit: None = None) -> Iterator[Workflow]: ...
+
+    @overload
+    def __call__(self, chunk_size: int, limit: None) -> Iterator[Workflow]: ...
+
+    def __call__(
+        self, chunk_size: int | None = None, limit: int | None = None
+    ) -> Iterator[Workflow] | Iterator[WorkflowList]:
+        """Iterate over workflows
+
+        Args:
+            chunk_size (int | None): The number of workflows to return in each chunk. Defaults to yielding one workflow at a time.
+            limit (int | None): Maximum number of workflows to return. Defaults to returning all items.
+
+        Returns:
+            Iterator[Workflow] | Iterator[WorkflowList]: Returns an iterator over workflows.
+
+        """
+        return self._list_generator(
+            method="GET", resource_cls=Workflow, list_cls=WorkflowList, limit=limit, chunk_size=chunk_size
+        )
+
+    def __iter__(self) -> Iterator[Workflow]:
+        """Iterate all over workflows"""
+        return self.__call__()
+
     def upsert(self, workflow: WorkflowUpsert, mode: Literal["replace"] = "replace") -> Workflow:
-        """`Create a workflow. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-versions/operation/CreateOrUpdateWorkflow>`_
+        """`Create a workflow. <https://api-docs.cognite.com/20230101/tag/Workflow-versions/operation/CreateOrUpdateWorkflow>`_
 
         Note this is an upsert endpoint, so if a workflow with the same external id already exists, it will be updated.
 
@@ -533,7 +583,6 @@ class WorkflowAPI(BetaWorkflowAPIClient):
                 >>> client = CogniteClient()
                 >>> res = client.workflows.upsert(WorkflowUpsert(external_id="my workflow", description="my workflow description"))
         """
-        self._warning.warn()
         if mode != "replace":
             raise ValueError("Only replace mode is supported for upserting workflows.")
 
@@ -544,7 +593,7 @@ class WorkflowAPI(BetaWorkflowAPIClient):
         return Workflow._load(response.json()["items"][0])
 
     def retrieve(self, external_id: str) -> Workflow | None:
-        """`Retrieve a workflow. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-versions/operation/CreateOrUpdateWorkflow>`_
+        """`Retrieve a workflow. <https://api-docs.cognite.com/20230101/tag/Workflow-versions/operation/CreateOrUpdateWorkflow>`_
 
         Args:
             external_id (str): Identifier for a Workflow. Must be unique for the project.
@@ -560,7 +609,6 @@ class WorkflowAPI(BetaWorkflowAPIClient):
                 >>> client = CogniteClient()
                 >>> res = client.workflows.retrieve("my workflow")
         """
-        self._warning.warn()
         try:
             response = self._get(url_path=f"{self._RESOURCE_PATH}/{quote(external_id, '')}")
         except CogniteAPIError as e:
@@ -584,18 +632,17 @@ class WorkflowAPI(BetaWorkflowAPIClient):
                 >>> client = CogniteClient()
                 >>> client.workflows.delete("my workflow")
         """
-        self._warning.warn()
         self._delete_multiple(
             identifiers=IdentifierSequence.load(external_ids=external_id),
             params={"ignoreUnknownIds": ignore_unknown_ids},
             wrap_ids=True,
         )
 
-    def list(self, limit: int = DEFAULT_LIMIT_READ) -> WorkflowList:
-        """`List all workflows in the project. <https://api-docs.cognite.com/20230101-beta/tag/Workflow-versions/operation/FetchAllWorkflows>`_
+    def list(self, limit: int | None = DEFAULT_LIMIT_READ) -> WorkflowList:
+        """`List all workflows in the project. <https://api-docs.cognite.com/20230101/tag/Workflow-versions/operation/FetchAllWorkflows>`_
 
         Args:
-            limit (int): Maximum number of results to return. Defaults to 25. Set to -1, float("inf") or None
+            limit (int | None): Maximum number of results to return. Defaults to 25. Set to -1, float("inf") or None
         Returns:
             WorkflowList: All workflows in the CDF project.
 
@@ -607,8 +654,6 @@ class WorkflowAPI(BetaWorkflowAPIClient):
                 >>> client = CogniteClient()
                 >>> res = client.workflows.list()
         """
-        self._warning.warn()
-
         return self._list(
             method="GET",
             resource_cls=Workflow,
