@@ -32,7 +32,7 @@ class DirectRelationReference:
         }
 
     @classmethod
-    def load(cls, data: dict | tuple[str, str]) -> DirectRelationReference:
+    def load(cls, data: dict | tuple[str, str] | DirectRelationReference) -> DirectRelationReference:
         if isinstance(data, dict):
             return cls(
                 space=data["space"],
@@ -40,6 +40,8 @@ class DirectRelationReference:
             )
         elif isinstance(data, tuple) and len(data) == 2:
             return cls(data[0], data[1])
+        elif isinstance(data, cls):
+            return data
         else:
             raise ValueError("Invalid data provided to load method. Must be dict or tuple with two elements.")
 
@@ -73,7 +75,7 @@ class PropertyType(CogniteObject, ABC):
         type_ = resource["type"]
         obj: Any
         if type_ == "text":
-            obj = Text(is_list=resource["list"], collation=resource["collation"])
+            obj = Text(is_list=resource["list"], collation=resource.get("collation", "ucs_basic"))
         elif type_ == "boolean":
             obj = Boolean(is_list=resource["list"])
         elif type_ == "float32":
@@ -99,7 +101,10 @@ class PropertyType(CogniteObject, ABC):
         elif type_ == "direct":
             obj = DirectRelation(
                 container=ContainerId.load(container) if (container := resource.get("container")) else None,
-                is_list=resource["list"],
+                # The PropertyTypes are used as both read and write objects. The `list` was added later
+                # in the API for DirectRelations. Thus, we need to set the default value to False
+                # to avoid breaking changes. When used as a read object, the `list` will always be present.
+                is_list=resource.get("list", False),
             )
         else:
             logger.warning(f"Unknown property type: {type_}")
