@@ -2,19 +2,21 @@ from __future__ import annotations
 
 import re
 import warnings
+from datetime import timezone
 from inspect import signature
 from itertools import chain
 from numbers import Integral
-from typing import TYPE_CHECKING, Any, Sequence
+from typing import TYPE_CHECKING, Any, Literal, Sequence
 
 from cognite.client.exceptions import CogniteImportError
 from cognite.client.utils._importing import local_import
 from cognite.client.utils._text import to_camel_case
-from cognite.client.utils._time import TIME_ATTRIBUTES
+from cognite.client.utils._time import TIME_ATTRIBUTES, ZoneInfo
 
 if TYPE_CHECKING:
     import pandas as pd
 
+    from cognite.client.data_classes import DatapointsArrayList, DatapointsList
     from cognite.client.data_classes._base import T_CogniteResource, T_CogniteResourceList
 
 
@@ -41,6 +43,48 @@ def pandas_major_version() -> int:
     from pandas import __version__
 
     return int(__version__.split(".")[0])
+
+
+def convert_tz_for_pandas(tz: str | timezone | ZoneInfo | None) -> str | timezone | None:
+    if tz is None or isinstance(tz, (str, timezone)):
+        return tz
+    if isinstance(tz, ZoneInfo):
+        # pandas is not happy about ZoneInfo :shrug:
+        if tz.key is not None:
+            return tz.key
+        raise ValueError("timezone of type ZoneInfo does not have the required 'key' attribute set")
+    raise ValueError(f"'timezone' not understood, expected one of: [None, datetime.timezone, ZoneInfo], got {type(tz)}")
+
+
+def concat_dps_dataframe_list(
+    dps_lst: DatapointsList | DatapointsArrayList,
+    column_names: Literal["id", "external_id"],
+    include_aggregate_name: bool,
+    include_granularity_name: bool,
+    include_status: bool,
+) -> pd.DataFrame:
+    pd = local_import("pandas")
+    dfs = [
+        dps.to_pandas(
+            column_names=column_names,
+            include_aggregate_name=include_aggregate_name,
+            include_granularity_name=include_granularity_name,
+            include_status=include_status,
+        )
+        for dps in dps_lst
+    ]
+    if not dfs:
+        return pd.DataFrame(index=pd.to_datetime([]))
+    timezones = set(dps.timezone for dps in dps_lst) - {None}
+    # If attempting to join naive & aware, pandas will raise (so we don't need to):
+    # TypeError: Cannot join tz-naive with tz-aware DatetimeIndex
+    if len(timezones) > 1:
+        warnings.warn(
+            f"When concatenating datapoints localized to different timezones ({sorted(map(str, timezones))}), the "
+            "final dataframe index (timestamps) will be a union of the UTC converted timestamps.",
+            UserWarning,
+        )
+    return concat_dataframes_with_nullable_int_cols(dfs)
 
 
 def notebook_display_with_fallback(inst: T_CogniteResource | T_CogniteResourceList, **kwargs: Any) -> str:
