@@ -3,6 +3,7 @@ from __future__ import annotations
 import json
 import math
 import time
+from datetime import date, datetime, timezone
 from typing import Any, ClassVar, cast
 
 import pytest
@@ -30,8 +31,11 @@ from cognite.client.data_classes.data_modeling import (
     NodeId,
     NodeList,
     NodeOrEdgeData,
+    PropertyOptions,
     SingleHopConnectionDefinition,
     Space,
+    TypedNode,
+    TypedNodeApply,
     View,
     ViewApply,
     ViewId,
@@ -121,16 +125,294 @@ def node_with_1_1_pressure_in_bar(
     return node
 
 
+class PrimitiveNullable(TypedNodeApply):
+    def __init__(
+        self,
+        space: str,
+        external_id: str,
+        text: str | None = None,
+        boolean: bool | None = None,
+        float32: float | None = None,
+        float64: float | None = None,
+        int32: int | None = None,
+        int64: int | None = None,
+        timestamp: datetime | None = None,
+        date: date | None = None,
+        json: dict | None = None,
+        direct: DirectRelationReference | None = None,
+        existing_version: int | None = None,
+        type: DirectRelationReference | tuple[str, str] | None = None,
+    ) -> None:
+        super().__init__(space=space, external_id=external_id, existing_version=existing_version, type=type)
+        self.text = text
+        self.boolean = boolean
+        self.float32 = float32
+        self.float64 = float64
+        self.int32 = int32
+        self.int64 = int64
+        self.timestamp = timestamp
+        self.date = date
+        self.json = json
+        self.direct = direct
+
+    @classmethod
+    def get_source(cls) -> ViewId:
+        return ViewId("IntegrationTestSpace", "PrimitiveNullable", "1")
+
+
+class PrimitiveNullableRead(TypedNode):
+    def __init__(
+        self,
+        space: str,
+        external_id: str,
+        version: int,
+        last_updated_time: int,
+        created_time: int,
+        text: str | None = None,
+        boolean: bool | None = None,
+        float32: float | None = None,
+        float64: float | None = None,
+        int32: int | None = None,
+        int64: int | None = None,
+        timestamp: datetime | None = None,
+        date: date | None = None,
+        json: dict | None = None,
+        direct: DirectRelationReference | None = None,
+        type: DirectRelationReference | tuple[str, str] | None = None,
+        deleted_time: int | None = None,
+    ) -> None:
+        super().__init__(
+            space=space,
+            external_id=external_id,
+            version=version,
+            last_updated_time=last_updated_time,
+            created_time=created_time,
+            type=type,
+            deleted_time=deleted_time,
+            properties=None,
+        )
+        self.text = text
+        self.boolean = boolean
+        self.float32 = float32
+        self.float64 = float64
+        self.int32 = int32
+        self.int64 = int64
+        self.timestamp = timestamp
+        self.date = date
+        self.json = json
+        self.direct = direct
+
+    def as_write(self) -> PrimitiveNullable:
+        return PrimitiveNullable(
+            space=self.space,
+            external_id=self.external_id,
+            text=self.text,
+            boolean=self.boolean,
+            float32=self.float32,
+            float64=self.float64,
+            int32=self.int32,
+            int64=self.int64,
+            timestamp=self.timestamp,
+            date=self.date,
+            json=self.json,
+            direct=self.direct,
+            existing_version=self.version,
+            type=self.type,
+        )
+
+    @classmethod
+    def get_source(cls) -> ViewId:
+        return ViewId("IntegrationTestSpace", "PrimitiveNullable", "1")
+
+
+class PrimitiveListed(TypedNodeApply):
+    def __init__(
+        self,
+        space: str,
+        external_id: str,
+        text: list[str] | None = None,
+        boolean: list[bool] | None = None,
+        float32: list[float] | None = None,
+        float64: list[float] | None = None,
+        int32: list[int] | None = None,
+        int64: list[int] | None = None,
+        timestamp: list[datetime] | None = None,
+        date: list[date] | None = None,
+        json: list[dict] | None = None,
+        direct: list[DirectRelationReference] | None = None,
+        existing_version: int | None = None,
+        type: DirectRelationReference | tuple[str, str] | None = None,
+    ) -> None:
+        super().__init__(space=space, external_id=external_id, existing_version=existing_version, type=type)
+        self.text = text
+        self.boolean = boolean
+        self.float32 = float32
+        self.float64 = float64
+        self.int32 = int32
+        self.int64 = int64
+        self.timestamp = timestamp
+        self.date = date
+        self.json = json
+        self.direct = direct
+
+    @classmethod
+    def get_source(cls) -> ViewId:
+        return ViewId("IntegrationTestSpace", "PrimitiveListed", "1")
+
+
+class PrimitiveListedRead(TypedNode):
+    def __init__(
+        self,
+        space: str,
+        external_id: str,
+        version: int,
+        last_updated_time: int,
+        created_time: int,
+        text: list[str] | None = None,
+        boolean: list[bool] | None = None,
+        float32: list[float] | None = None,
+        float64: list[float] | None = None,
+        int32: list[int] | None = None,
+        int64: list[int] | None = None,
+        timestamp: list[datetime] | None = None,
+        date: list[date] | None = None,
+        json: list[dict] | None = None,
+        direct: list[DirectRelationReference] | None = None,
+        type: DirectRelationReference | None = None,
+        deleted_time: int | None = None,
+    ) -> None:
+        super().__init__(
+            space=space,
+            external_id=external_id,
+            version=version,
+            last_updated_time=last_updated_time,
+            created_time=created_time,
+            type=type,
+            deleted_time=deleted_time,
+            properties=None,
+        )
+        self.text = text
+        self.boolean = boolean
+        self.float32 = float32
+        self.float64 = float64
+        self.int32 = int32
+        self.int64 = int64
+        self.timestamp = timestamp
+        self.date = date
+        self.json = json
+        self.direct = direct
+
+    @classmethod
+    def get_source(cls) -> ViewId:
+        return ViewId("IntegrationTestSpace", "PrimitiveListed", "1")
+
+    def as_write(self) -> PrimitiveListed:
+        return PrimitiveListed(
+            space=self.space,
+            external_id=self.external_id,
+            text=self.text,
+            boolean=self.boolean,
+            float32=self.float32,
+            float64=self.float64,
+            int32=self.int32,
+            int64=self.int64,
+            timestamp=self.timestamp,
+            date=self.date,
+            json=self.json,
+            direct=self.direct,
+            existing_version=self.version,
+            type=self.type,
+        )
+
+
+class Person(TypedNodeApply):
+    birth_year = PropertyOptions(identifier="birthYear")
+
+    def __init__(
+        self,
+        space: str,
+        external_id: str,
+        name: str,
+        birth_year: int | None = None,
+        existing_version: int | None = None,
+        type: DirectRelationReference | tuple[str, str] | None = None,
+    ):
+        super().__init__(space=space, external_id=external_id, existing_version=existing_version, type=type)
+        self.name = name
+        self.birth_year = birth_year
+
+    @classmethod
+    def get_source(cls) -> ViewId:
+        return ViewId("IntegrationTestSpace", "Person", "37ce1494b83df2")
+
+
+class PersonRead(TypedNode):
+    birth_year = PropertyOptions(identifier="birthYear")
+
+    def __init__(
+        self,
+        space: str,
+        external_id: str,
+        version: int,
+        last_updated_time: int,
+        created_time: int,
+        name: str,
+        birth_year: int | None = None,
+        type: DirectRelationReference | None = None,
+        deleted_time: int | None = None,
+    ):
+        super().__init__(
+            space=space,
+            external_id=external_id,
+            version=version,
+            last_updated_time=last_updated_time,
+            created_time=created_time,
+            type=type,
+            deleted_time=deleted_time,
+            properties=None,
+        )
+        self.name = name
+        self.birth_year = birth_year
+
+    @classmethod
+    def get_source(cls) -> ViewId:
+        return ViewId("IntegrationTestSpace", "Person", "37ce1494b83df2")
+
+    def as_write(self) -> Person:
+        return Person(
+            space=self.space,
+            external_id=self.external_id,
+            name=self.name,
+            birth_year=self.birth_year,
+            existing_version=self.version,
+            type=self.type,
+        )
+
+
 class TestInstancesAPI:
     def test_list_nodes(self, cognite_client: CogniteClient, movie_nodes: NodeList) -> None:
-        listed_nodes = cognite_client.data_modeling.instances.list(limit=-1, instance_type="node")
+        is_prefix = filters.Or(
+            *[
+                filters.Prefix(["node", "externalId"], prefix)
+                for prefix in {n.external_id.split(":", maxsplit=1)[0] for n in movie_nodes}
+            ]
+        )
+
+        listed_nodes = cognite_client.data_modeling.instances.list(limit=-1, instance_type="node", filter=is_prefix)
 
         movie_node_ids = set(movie_nodes.as_ids())
         assert movie_node_ids
         assert movie_node_ids <= set(listed_nodes.as_ids())
 
     def test_list_edges(self, cognite_client: CogniteClient, movie_edges: EdgeList) -> None:
-        listed_edges = cognite_client.data_modeling.instances.list(limit=-1, instance_type="edge")
+        is_prefix = filters.Or(
+            *[
+                filters.Prefix(["edge", "externalId"], prefix)
+                for prefix in {e.external_id.split(":", maxsplit=1)[0] for e in movie_edges}
+            ]
+        )
+
+        listed_edges = cognite_client.data_modeling.instances.list(limit=-1, instance_type="edge", filter=is_prefix)
         assert set(movie_edges.as_ids()) <= set(listed_edges.as_ids())
 
     def test_list_nodes_with_properties(self, cognite_client: CogniteClient, person_view: View) -> None:
@@ -337,6 +619,14 @@ class TestInstancesAPI:
         )
         assert set(retrieved.nodes.as_ids()) == set(movie_nodes.as_ids())
         assert set(retrieved.edges.as_ids()) == set(movie_edges.as_ids())
+
+    def test_retrieve_nodes(self, cognite_client: CogniteClient, movie_nodes: NodeList) -> None:
+        retrieved = cognite_client.data_modeling.instances.retrieve_nodes(movie_nodes.as_ids())
+        assert set(retrieved.as_ids()) == set(movie_nodes.as_ids())
+
+    def test_retrieve_edges(self, cognite_client: CogniteClient, movie_edges: EdgeList) -> None:
+        retrieved = cognite_client.data_modeling.instances.retrieve_edges(movie_edges.as_ids())
+        assert set(retrieved.as_ids()) == set(movie_edges.as_ids())
 
     def test_retrieve_multiple_with_missing(self, cognite_client: CogniteClient, movie_nodes: NodeList) -> None:
         ids_without_missing = movie_nodes.as_ids()
@@ -595,7 +885,7 @@ class TestInstancesAPI:
 
         retrieved = cognite_client.data_modeling.instances.retrieve(node.as_id(), sources=[source])
         assert retrieved.nodes
-        assert math.isclose(retrieved.nodes[0]["pressure"], 1.1 * 1e5)
+        assert math.isclose(cast(float, retrieved.nodes[0]["pressure"]), 1.1 * 1e5)
 
     def test_list_in_units(
         self, cognite_client: CogniteClient, node_with_1_1_pressure_in_bar: NodeApply, unit_view: View
@@ -606,7 +896,7 @@ class TestInstancesAPI:
 
         assert listed
         assert len(listed) == 1
-        assert math.isclose(listed[0]["pressure"], 1.1 * 1e5)
+        assert math.isclose(cast(float, listed[0]["pressure"]), 1.1 * 1e5)
 
     def test_search_in_units(
         self, cognite_client: CogniteClient, node_with_1_1_pressure_in_bar: NodeApply, unit_view: View
@@ -620,7 +910,7 @@ class TestInstancesAPI:
 
         assert searched
         assert len(searched) == 1
-        assert math.isclose(searched[0]["pressure"], 1.1 * 1e5)
+        assert math.isclose(cast(float, searched[0]["pressure"]), 1.1 * 1e5)
 
     def test_aggregate_in_units(
         self, cognite_client: CogniteClient, node_with_1_1_pressure_in_bar: NodeApply, unit_view: View
@@ -653,6 +943,107 @@ class TestInstancesAPI:
         assert queried
         assert len(queried["nodes"]) == 1
         assert math.isclose(queried["nodes"][0]["pressure"], 1.1 * 1e5)
+
+    @pytest.mark.usefixtures("primitive_nullable_view")
+    def test_write_typed_node(self, cognite_client: CogniteClient, integration_test_space: Space) -> None:
+        space = integration_test_space.space
+        external_id = "node_test_write_read_custom_properties"
+        primitive = PrimitiveNullable(
+            space=space,
+            external_id=external_id,
+            text="text",
+            boolean=True,
+            float32=1.1,
+            float64=1.1,
+            int32=1,
+            int64=1,
+            # Replacing microseconds with 0 to avoid comparison issues in the asserting below
+            # as server only stores milliseconds
+            timestamp=datetime.now(timezone.utc).replace(microsecond=0),
+            date=date.today(),
+            json={"key": "value", "nested": {"key": "value"}},
+            direct=DirectRelationReference(space, external_id),
+        )
+        try:
+            created = cognite_client.data_modeling.instances.apply(primitive)
+            assert len(created.nodes) == 1
+            assert created.nodes[0].external_id == external_id
+
+            retrieved = cognite_client.data_modeling.instances.retrieve_nodes(
+                primitive.as_id(), node_cls=PrimitiveNullableRead
+            )
+            assert isinstance(retrieved, PrimitiveNullableRead)
+            assert retrieved.text == "text"
+        finally:
+            cognite_client.data_modeling.instances.delete(primitive.as_id())
+
+    @pytest.mark.usefixtures("primitive_nullable_listed_view")
+    def test_write_typed_node_listed_properties(
+        self, cognite_client: CogniteClient, integration_test_space: Space
+    ) -> None:
+        space = integration_test_space.space
+        external_id = "node_test_write_read_custom_properties_listable"
+        primitive_listed = PrimitiveListed(
+            space=space,
+            external_id=external_id,
+            text=["text"],
+            boolean=[True],
+            float32=[1.1],
+            float64=[1.1],
+            int32=[1],
+            int64=[1],
+            # Replacing microseconds with 0 to avoid comparison issues in the asserting below
+            timestamp=[datetime.now(timezone.utc).replace(microsecond=0)],
+            date=[date.today()],
+            json=[{"key": "value", "nested": {"key": "value"}}],
+            direct=[DirectRelationReference(space, external_id)],
+        )
+        try:
+            created = cognite_client.data_modeling.instances.apply(primitive_listed)
+            assert len(created.nodes) == 1
+            assert created.nodes[0].external_id == external_id
+
+            retrieved = cognite_client.data_modeling.instances.retrieve_nodes(
+                primitive_listed.as_id(), node_cls=PrimitiveListedRead
+            )
+            assert isinstance(retrieved, PrimitiveListedRead)
+            assert retrieved.text == ["text"]
+        finally:
+            cognite_client.data_modeling.instances.delete(primitive_listed.as_id())
+
+    @pytest.mark.usefixtures("person_view")
+    def test_write_type_node_instance_property_descriptor(
+        self, cognite_client: CogniteClient, integration_test_space: Space
+    ) -> None:
+        space = integration_test_space.space
+        external_id = "node_test_write_read_instance_property_descriptor"
+        person = Person(space=space, external_id=external_id, name="John Doe", birth_year=1980)
+
+        try:
+            created = cognite_client.data_modeling.instances.apply(person)
+            assert len(created.nodes) == 1
+            assert created.nodes[0].external_id == external_id
+
+            retrieved = cognite_client.data_modeling.instances.retrieve_nodes(person.as_id(), node_cls=PersonRead)
+            assert isinstance(retrieved, PersonRead)
+            assert retrieved.name == "John Doe"
+            assert retrieved.birth_year == 1980
+        finally:
+            cognite_client.data_modeling.instances.delete(person.as_id())
+
+    def test_list_person_types(self, cognite_client: CogniteClient) -> None:
+        persons = cognite_client.data_modeling.instances.list(PersonRead, limit=10)
+
+        assert len(persons) > 0
+        assert all(isinstance(person, PersonRead) for person in persons)
+
+    def test_search_person(self, cognite_client: CogniteClient) -> None:
+        persons = cognite_client.data_modeling.instances.search(
+            PersonRead.get_source(), query="Quentin", instance_type=PersonRead, limit=10
+        )
+
+        assert len(persons) > 0
+        assert all(isinstance(person, PersonRead) for person in persons)
 
 
 class TestInstancesSync:
