@@ -1,6 +1,7 @@
 from __future__ import annotations
 
-from typing import Any
+from datetime import date, datetime
+from typing import Any, List, Union, cast
 
 import pytest
 
@@ -14,10 +15,11 @@ from cognite.client.data_classes.data_modeling import (
     NodeApply,
     NodeId,
     NodeList,
+    NodeListWithCursor,
     NodeOrEdgeData,
     ViewId,
 )
-from cognite.client.data_classes.data_modeling.instances import Instance
+from cognite.client.data_classes.data_modeling.instances import EdgeListWithCursor, Instance
 
 
 class TestEdgeApply:
@@ -78,8 +80,16 @@ class TestNodeApply:
                     properties=dict(
                         name="Integration test",
                         scenario="Integration test",
-                        start_time="2021-01-01T00:00:00",
-                        end_time="2021-01-01T00:00:00",
+                        date=date(2024, 6, 18),
+                        start_time=datetime.fromisoformat("2021-01-01T00:00:00"),
+                        end_time=datetime.fromisoformat("2021-01-01T00:00:00"),
+                        other_nodes=cast(
+                            List[Union[NodeId, DirectRelationReference]],
+                            [
+                                DirectRelationReference("space", "external_id"),
+                                NodeId("space", "external_id2"),
+                            ],
+                        ),
                         cut_files=["shop:cut_file:1"],
                         bid="shop:bid_matrix:8",
                         bid_history=["shop:bid_matrix:9"],
@@ -108,11 +118,16 @@ class TestNodeApply:
                             "space": "IntegrationTestsImmutable",
                         },
                         "cut_files": ["shop:cut_file:1"],
-                        "end_time": "2021-01-01T00:00:00",
+                        "end_time": "2021-01-01T00:00:00.000",
                         "name": "Integration test",
                         "runStatus": "Running",
                         "scenario": "Integration test",
-                        "start_time": "2021-01-01T00:00:00",
+                        "start_time": "2021-01-01T00:00:00.000",
+                        "date": "2024-06-18",
+                        "other_nodes": [
+                            {"externalId": "external_id", "space": "space"},
+                            {"externalId": "external_id2", "space": "space"},
+                        ],
                     },
                     "source": {"externalId": "Case", "space": "IntegrationTestsImmutable", "type": "container"},
                 }
@@ -145,6 +160,118 @@ class TestNode:
             "type": {"externalId": "someType", "space": "someSpace"},
             "version": 1,
         }
+
+
+class TestNodeListWithCursor:
+    def test_extend(self) -> None:
+        default_args: dict[str, Any] = dict(
+            version=1, last_updated_time=0, created_time=0, deleted_time=None, properties=None, type=None
+        )
+        nodes = NodeListWithCursor[Node](
+            [
+                Node(space="space", external_id="node1", **default_args),
+                Node(space="space", external_id="node2", **default_args),
+            ],
+            cursor="original_cursor",
+        )
+
+        nodes.extend(
+            NodeListWithCursor[Node](
+                [
+                    Node(space="space", external_id="node3", **default_args),
+                    Node(space="space", external_id="node4", **default_args),
+                ],
+                cursor="next_cursor",
+            ),
+        )
+        assert len(nodes) == 4
+        assert nodes.cursor == "next_cursor"
+
+    def test_extend_with_duplicates(self) -> None:
+        default_args: dict[str, Any] = dict(
+            version=1, last_updated_time=0, created_time=0, deleted_time=None, properties=None, type=None
+        )
+        nodes = NodeListWithCursor[Node](
+            [
+                Node(space="space", external_id="node1", **default_args),
+                Node(space="space", external_id="node2", **default_args),
+            ],
+            cursor="original_cursor",
+        )
+
+        with pytest.raises(ValueError):
+            nodes.extend(
+                NodeListWithCursor[Node](
+                    [
+                        Node(space="space", external_id="node2", **default_args),
+                        Node(space="space", external_id="node3", **default_args),
+                    ],
+                    cursor="next_cursor",
+                ),
+            )
+
+
+class TestEdgeListWithCursor:
+    def test_extend(self) -> None:
+        default_args: dict[str, Any] = dict(
+            start_node=DirectRelationReference("space", "node1"),
+            end_node=DirectRelationReference("space", "node2"),
+            version=1,
+            last_updated_time=0,
+            created_time=0,
+            deleted_time=None,
+            properties=None,
+            type=DirectRelationReference("space", "type"),
+        )
+        edges = EdgeListWithCursor(
+            [
+                Edge(space="space", external_id="edge1", **default_args),
+                Edge(space="space", external_id="edge2", **default_args),
+            ],
+            cursor="original_cursor",
+        )
+
+        edges.extend(
+            EdgeListWithCursor(
+                [
+                    Edge(space="space", external_id="edge3", **default_args),
+                    Edge(space="space", external_id="edge4", **default_args),
+                ],
+                cursor="next_cursor",
+            ),
+        )
+        assert len(edges) == 4
+        assert edges.cursor == "next_cursor"
+
+    def test_extend_with_duplicates(self) -> None:
+        default_args: dict[str, Any] = dict(
+            start_node=DirectRelationReference("space", "node1"),
+            end_node=DirectRelationReference("space", "node2"),
+            version=1,
+            last_updated_time=0,
+            created_time=0,
+            deleted_time=None,
+            properties=None,
+            type=DirectRelationReference("space", "type"),
+        )
+        edges = EdgeListWithCursor(
+            [
+                Edge(space="space", external_id="edge1", **default_args),
+                Edge(space="space", external_id="edge2", **default_args),
+            ],
+            cursor="original_cursor",
+        )
+
+        with pytest.raises(ValueError):
+            edges.extend(
+                EdgeListWithCursor(
+                    [
+                        Edge(space="space", external_id="edge2", **default_args),
+                        Edge(space="space", external_id="edge3", **default_args),
+                    ],
+                    cursor="next_cursor",
+                ),
+            )
 
 
 @pytest.fixture
