@@ -99,7 +99,7 @@ class TestHTTPClient:
         )
 
         with pytest.raises(CogniteReadTimeout):
-            c.request("GET", "bla")
+            c.request("GET", "bla", accept="application/json")
 
         assert retry_tracker.total == DEFAULT_CONFIG.max_retries_read
         assert retry_tracker.read == DEFAULT_CONFIG.max_retries_read
@@ -123,7 +123,7 @@ class TestHTTPClient:
         )
 
         with pytest.raises(CogniteConnectionError):
-            c.request("GET", "bla")
+            c.request("GET", "bla", accept="application/json")
 
         assert retry_tracker.total == DEFAULT_CONFIG.max_retries_connect
         assert retry_tracker.read == 0
@@ -148,7 +148,7 @@ class TestHTTPClient:
         )
 
         with pytest.raises(CogniteConnectionRefused):
-            c.request("GET", "bla")
+            c.request("GET", "bla", accept="application/json")
 
         assert retry_tracker.total == DEFAULT_CONFIG.max_retries_connect
 
@@ -163,7 +163,7 @@ class TestHTTPClient:
             session=MagicMock(request=MagicMock(return_value=MagicMock(status_code=429))),
         )
 
-        res = c.request("GET", "bla")
+        res = c.request("GET", "bla", accept="application/json")
         assert res.status_code == 429
 
         assert retry_tracker.total == DEFAULT_CONFIG.max_retries_status
