@@ -3,7 +3,6 @@ from operator import attrgetter
 import pytest
 
 from cognite.client import ClientConfig, CogniteClient
-from cognite.client._api.workflows import BetaWorkflowAPIClient
 from cognite.client._api_client import APIClient
 from cognite.client.credentials import Token
 from cognite.client.testing import CogniteClientMock, monkeypatch_cognite_client
@@ -13,7 +12,7 @@ from tests.utils import all_mock_children, all_subclasses
 def test_ensure_all_apis_are_available_on_cognite_mock():
     mocked_apis = all_mock_children(CogniteClientMock())
     available = {v.__class__ for v in mocked_apis.values()}
-    expected = set(all_subclasses(APIClient)) - {BetaWorkflowAPIClient}
+    expected = set(all_subclasses(APIClient))
     # Any new APIs that have not been added to CogniteClientMock?
     assert not expected.difference(available), f"Missing APIs: {expected.difference(available)}"
     # Any removed APIs that are still available on CogniteClientMock?
