@@ -10,9 +10,12 @@ import math
 import os
 import random
 import string
+import sys
 import typing
 from collections import Counter
 from contextlib import contextmanager
+from datetime import timedelta, timezone
+from pathlib import Path
 from typing import TYPE_CHECKING, Any, Literal, Mapping, TypeVar, cast, get_args, get_origin, get_type_hints
 
 from cognite.client import CogniteClient
@@ -31,11 +34,14 @@ from cognite.client.data_classes import (
 )
 from cognite.client.data_classes._base import CogniteResourceList, Geometry
 from cognite.client.data_classes.aggregations import Buckets
+from cognite.client.data_classes.capabilities import Capability, LegacyCapability, UnknownAcl
+from cognite.client.data_classes.data_modeling import TypedEdge, TypedEdgeApply, TypedNode, TypedNodeApply
 from cognite.client.data_classes.data_modeling.query import NodeResultSetExpression, Query
 from cognite.client.data_classes.datapoints import _INT_AGGREGATES, ALL_SORTED_DP_AGGS, Datapoints, DatapointsArray
 from cognite.client.data_classes.filters import Filter
 from cognite.client.data_classes.transformations.notifications import TransformationNotificationWrite
 from cognite.client.data_classes.transformations.schedules import TransformationScheduleWrite
+from cognite.client.data_classes.transformations.schema import TransformationSchemaUnknownType
 from cognite.client.data_classes.workflows import (
     FunctionTaskOutput,
     FunctionTaskParameters,
@@ -50,7 +56,15 @@ from cognite.client.utils._text import random_string, to_snake_case
 if TYPE_CHECKING:
     import pandas
 
+REPO_ROOT = Path(__file__).resolve().parent.parent
+
 T_Type = TypeVar("T_Type", bound=type)
+
+UNION_TYPES = {typing.Union}
+if sys.version_info >= (3, 10):
+    from types import UnionType
+
+    UNION_TYPES.add(UnionType)
 
 
 def all_subclasses(base: T_Type) -> list[T_Type]:
@@ -70,7 +84,11 @@ def all_concrete_subclasses(base: T_Type) -> list[T_Type]:
     return [
         sub
         for sub in all_subclasses(base)
-        if all(base is not abc.ABC for base in sub.__bases__) and not inspect.isabstract(sub)
+        if all(base is not abc.ABC for base in sub.__bases__)
+        and not inspect.isabstract(sub)
+        # The FakeCogniteResourceGenerator does not support descriptors, so we exclude the Typed classes
+        # as these use the PropertyOptions descriptor.
+        and all(parent not in {TypedNodeApply, TypedEdgeApply, TypedEdge, TypedNode} for parent in sub.__mro__)
     ]
 
 
@@ -351,6 +369,8 @@ class FakeCogniteResourceGenerator:
             else:
                 for raw in ["value", "status_code", "status_symbol"]:
                     keyword_arguments.pop(raw, None)
+        elif resource_cls is TransformationSchemaUnknownType:
+            keyword_arguments["raw_schema"]["type"] = "unknown"
         elif resource_cls is SequenceRows:
             # All row values must match the number of columns
             # Reducing to one column, and one value for each row
@@ -395,6 +415,9 @@ class FakeCogniteResourceGenerator:
         elif resource_cls is TransformationScheduleWrite:
             # TransformationScheduleWrite requires either id or external_id
             keyword_arguments.pop("id", None)
+            if skip_defaulted_args:
+                # At least external_id or id must be set
+                keyword_arguments["external_id"] = "my_schedule"
         elif resource_cls is TransformationNotificationWrite:
             # TransformationNotificationWrite requires either transformation_id or transformation_external_id
             if skip_defaulted_args:
@@ -406,10 +429,14 @@ class FakeCogniteResourceGenerator:
             keyword_arguments["through"] = [keyword_arguments["through"][0], "my_view/v1", "a_property"]
         elif resource_cls is Buckets:
             keyword_arguments = {"items": [{"start": 1, "count": 1}]}
+        elif resource_cls is timezone:
+            positional_arguments.append(timedelta(hours=self._random.randint(-3, 3)))
 
         return resource_cls(*positional_arguments, **keyword_arguments)
 
     def create_value(self, type_: Any, var_name: str | None = None) -> Any:
+        import numpy as np
+
         if isinstance(type_, typing.ForwardRef):
             type_ = type_._evaluate(globals(), self._type_checking())
 
@@ -442,6 +469,10 @@ class FakeCogniteResourceGenerator:
                 implementations.remove(filters.GeoJSONWithin)
                 implementations.remove(filters.GeoJSONDisjoint)
                 implementations.remove(filters.GeoJSONIntersects)
+            elif type_ is Capability:
+                implementations.remove(UnknownAcl)
+                if LegacyCapability in implementations:
+                    implementations.remove(LegacyCapability)
             if type_ is WorkflowTaskOutput:
                 # For Workflow Output has to match the input type
                 selected = FunctionTaskOutput
@@ -461,9 +492,8 @@ class FakeCogniteResourceGenerator:
 
         container_type = get_origin(type_)
         is_container = container_type is not None
-        if not is_container:
+        if not is_container or container_type is np.ndarray:  # looks weird, but 3.8 and 3.12 type compat. issue
             # Handle numpy types
-            import numpy as np
             from numpy.typing import NDArray
 
             if type_ == NDArray[np.float64]:
@@ -482,7 +512,7 @@ class FakeCogniteResourceGenerator:
         # Handle containers
         args = get_args(type_)
         first_not_none = next((arg for arg in args if arg is not None), None)
-        if container_type is typing.Union:
+        if container_type in UNION_TYPES:
             return self.create_value(first_not_none)
         elif container_type is typing.Literal:
             return self._random.choice(args)
